safari-600.7-branch/LayoutTests/fast/dom/Window/open-invalid-url-expected.txt

CONSOLE MESSAGE: Unable to open a window with invalid URL '/'.

ALERT: PASS
